# api/docs/swagger.yaml
definitions:
  ASError:
    properties:
      details:
        type: string
      errors_map:
        $ref: '#/definitions/validator.ValidationErrorsTranslations'
      internal_status:
        type: string
      message:
        type: string
    required:
    - details
    - internal_status
    - message
    type: object
...
  CurrentUser:
    properties:
      avatar_url:
        type: string
        x-nullable: true
      email:
        type: string
      id:
        type: string
      name:
        type: string
      timezone:
        type: string
    required:
    - avatar_url
    - email
    - id
    - name
    - timezone
    type: object
  PaginatedData-User:
    properties:
      results:
        items:
          $ref: '#/definitions/User'
        type: array
      total_count:
        type: integer
    required:
    - results
    - total_count
    type: object
  Project:
    properties:
      avatar_url:
        type: string
        x-nullable: true
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      plan:
        type: string
      slug:
        type: string
      updated_at:
        type: string
        x-nullable: true
    required:
    - avatar_url
    - created_at
    - id
    - name
    - plan
    - slug
    - updated_at
    type: object
...
  ProjectTeamMembership:
    properties:
      archived_at:
        type: string
        x-nullable: true
      created_at:
        type: string
      id:
        type: string
      position:
        type: string
        x-nullable: true
      team_id:
        type: string
      user_id:
        type: string
    required:
    - archived_at
    - created_at
    - id
    - position
    - team_id
    - user_id
    type: object
...
  User:
    properties:
      approved_at:
        type: string
        x-nullable: true
      archived_at:
        type: string
        x-nullable: true
      avatar_url:
        type: string
        x-nullable: true
      created_at:
        type: string
      email:
        format: email
        type: string
      id:
        type: string
      name:
        type: string
      referrer:
        type: string
        x-nullable: true
      updated_at:
        type: string
        x-nullable: true
      verified_at:
        type: string
        x-nullable: true
    required:
    - approved_at
    - archived_at
    - avatar_url
    - created_at
    - email
    - id
    - name
    - referrer
    - updated_at
    - verified_at
    type: object
...
  validator.ValidationErrorsTranslations:
    additionalProperties:
      type: string
    type: object
host: api.asyncstatus.com
info:
  contact: {}
  title: AsyncStatus API
  version: "0.1"
paths:
...
  /auth/session/user:
    get:
      consumes:
      - application/json
      operationId: get-current-user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CurrentUser'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ASError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ASError'
      summary: returns current user from session, only call it when you sure that
        user is authenticated (e.g. you fetch get-session in your middleware)
      tags:
      - auth
...
  /users:
    get:
      consumes:
      - application/json
      operationId: get-user
      parameters:
      - format: email
        in: query
        name: email
        type: string
      - in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ASError'
      summary: get user
      tags:
      - user
  /users/{user_id}/projects:
    get:
      consumes:
      - application/json
      operationId: get-user-projects
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Project'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ASError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ASError'
      summary: get user projects
      tags:
      - user
  /users/{user_id}/projects/{project_slug}/memberships/{team_slug}:
    get:
      consumes:
      - application/json
      operationId: get-user-team-membership
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Team slug
        in: path
        name: team_slug
        required: true
        type: string
      - description: Project slug
        in: path
        name: project_slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ProjectTeamMembership'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ASError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ASError'
      summary: get user team membership
      tags:
      - user
  /users/list:
    get:
      consumes:
      - application/json
      operationId: list-users
      parameters:
      - default: 50
        in: query
        minimum: 1
        name: limit
        type: integer
      - default: 0
        in: query
        minimum: 0
        name: offset
        type: integer
      - in: query
        name: orderby
        type: string
      - default: DESC
        in: query
        name: orderdirection
        type: string
      - description: Stringified JSON of `[]PaginatedBodySearch` type
        in: query
        name: search
        type: string
      - default: list
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PaginatedData-User'
      summary: list users
      tags:
      - user
swagger: "2.0"